requirements.txt
Flask==2.2.2
PyYAML==6.0.1
requests==2.28.2
waitress==3.0.0
Werkzeug==2.2.2